+++ noodles-sam/src/record/data/field/value/base_modifications/group.rs
@@ -1,3 +1,5 @@
+//! Base modifications group.
+
 pub mod modification;
 mod status;
 mod strand;